[FD-7495] Skip FreshID user creation for custserv@freshdesk.com (#1740)
JIRA Id(s) : FD-7495

Description :

Why is this change necessary?
custserv@freshdesk.com is not in use. So, we need add a check in Freshdesk to not to create this user in FreshID to avoid timeouts.
How does it address the issue?
Do not create user in FreshID for custserv email.

+++ app/models/user.rb
@@ -1111,7 +1111,7 @@ class User < ActiveRecord::Base
   end
 
   def create_freshid_user
-    return unless freshid_enabled_and_agent?
+    return if freshid_disabled_or_customer?
     Rails.logger.info "FRESHID Creating user :: a=#{self.account_id}, u=#{self.id}, email=#{self.email}"
     reset_freshid_user if email_changed?
     freshid_user = Freshid::User.create(user_attributes_for_freshid)
@@ -1129,7 +1129,7 @@ class User < ActiveRecord::Base
   end
 
   def destroy_freshid_user
-    if freshid_enabled_account? && freshid_authorization.present?
+    if freshid_enabled_account? && email_allowed_in_freshid? && freshid_authorization.present?
       remove_freshid_user
       freshid_authorization.destroy
       self.password_salt = self.crypted_password = nil
@@ -1178,11 +1178,15 @@ class User < ActiveRecord::Base
     end
 
     def freshid_enabled_and_agent?
-      agent? && freshid_enabled_account?
+      agent? && freshid_enabled_account? && email_allowed_in_freshid?
     end
 
-    def freshid_disabled_and_customer?
+    def freshid_disabled_or_customer?
       !freshid_enabled_and_agent?
+    end
+
+    def email_allowed_in_freshid?
+      !FRESHID_IGNORED_EMAIL_IDS.include?(self.email)
     end
 
     def valid_freshid_login?(incoming_password)
